Revert split dataeast/deco156.cpp into separate drivers and rename to dataeast/hvysmsh.cpp and updates:
- Split driver state via hardware configurations
- Fix namings
- Reduce duplicates
- Fix metadata via in-game version information screen and title screen
- Fix notes for version information screen

DECO156 decrypt routine source code file name is keeped to dataeast/deco156.cpp
